Correct state detection and handling
NoiseAlert doesn't try to remember service status but instead reads the static
flag in service class. This required to have the service running in the same
process as the main activity, but separated process was unnecessary overhead
anyways. We also fixed UI: hide test button when Test running and keep status
message on top of main activity up-to-date.